+++ handcrafted-haven/src/app/products/[productId]/edit/not-found.tsx
@@ -1,5 +1,12 @@
 import Link from 'next/link';
 
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Not Found",
+  description: "Could not find the requested product.",
+};
+
 
 export default function NotFound() {
   return (
