+++ handcrafted-haven/src/components/products/edit-form.tsx
@@ -196,7 +196,7 @@ export default function EditProductForm({product}: {product: ProductForm}) {
     
                     <Link
 
-                        href="/products/my-products"
+                        href="/profile"
                         className="flex h-10 items-center rounded-lg bg-gray-100 px-4 text-sm font-medium text-gray-600 transition-colors hover:bg-gray-200"
                     >
                         Cancel
